+++ client/components/about/AboutTextSection.tsx
@@ -1,0 +1,24 @@
+const AboutTextSection = () => (
+  <div className="w-full rounded-2xl border border-black bg-gray-200 p-5 md:w-2/5">
+    <p className="mb-4 text-black">
+      The Society of Asian Scientists and Engineers (SASE) at UF was established
+      in 2010 as a student organization dedicated to the advancement of Asian
+      heritage scientists and engineers. Above all, SASE is a welcoming
+      community of driven individuals who celebrate culture and diversity,
+      cultivating a unique balance between professional development and a
+      supportive, fun community.
+    </p>
+    <p className="text-black">
+      Through our general body meetings, socials, workshops, and internal
+      mentorship program, we aim to build our members’ relationships, skills,
+      and knowledge that will enable them to succeed. We also offer an
+      intramural program (SASE Sports), an engineering team (SET), and a
+      leadership development program (SASE Interns). UF SASE continues to
+      provide more development opportunities to our members year after year, all
+      while staying as welcoming as ever! Even if you’re not a STEM major, or of
+      Asian heritage, SASE welcomes you with open arms.
+    </p>
+  </div>
+);
+
+export default AboutTextSection;
