+++ client/components/about/HeaderSection.tsx
@@ -1,0 +1,11 @@
+const HeaderSection = () => (
+  <header className="mb-10 text-center">
+    <h1 className="text-4xl font-bold text-black">About Us</h1>
+    <p className="mt-4 text-xl text-black">
+      "We help shape skills and provide knowledge that will help our members to
+      succeed in the professional world"
+    </p>
+  </header>
+);
+
+export default HeaderSection;
