+++ client/components/about/HistorySection.tsx
@@ -1,0 +1,31 @@
+const HistorySection = () => (
+  <section className="mb-12">
+    <h2 className="mb-4 text-2xl font-semibold text-black">Our History</h2>
+    <p className="mb-4 text-black">
+      Since being founded during the summer of 2010, the University of Florida
+      chapter has had an abundant effort put forth for the development of our
+      members.
+    </p>
+    <p className="mb-4 text-black">
+      This development is centered on five core values of our mission statement:
+      leadership, professionalism, service, academics, and diversity. Our
+      meetings and events are catered to ensure that not only our mission is
+      being fulfilled, but also that a community of support among Asian heritage
+      students is being built. We pride ourselves on our measurable successes
+      but could not have done so without the help of our members, faculty, and
+      sponsors.
+    </p>
+    <div className="mt-5 text-center">
+      <a
+        href="https://ufsase.com/wp-content/uploads/2024/03/2024-2025-SASE-Constitution.docx.pdf"
+        className="mt-6 inline-block transform rounded-full bg-[#0f6cb6] px-6 py-2 text-white transition duration-300 hover:scale-105 hover:duration-300"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        2024-2025 SASE Constitution
+      </a>
+    </div>
+  </section>
+);
+
+export default HistorySection;
